refactor(whatsrust): extract presence wrappers
